# image_server/generated-client/types.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

export type ClientOptions = {
    baseUrl: 'http://localhost:3007' | (string & {});
};

export type ImageName = {
    /**
     * Image filename
     */
    name: string;
};

export type ImageDeleted = {
    /**
     * Name of the deleted image
     */
    deleted: string;
};

export type Error = {
    /**
     * Error message
     */
    error: string;
};

export type ListImagesData = {
    body?: never;
    path?: never;
    query?: never;
    url: '/images';
};

export type ListImagesErrors = {
    /**
     * Server error
     */
    500: Error;
};

export type ListImagesError = ListImagesErrors[keyof ListImagesErrors];

export type ListImagesResponses = {
    /**
     * Array of image names
     */
    200: Array<ImageName>;
};

export type ListImagesResponse = ListImagesResponses[keyof ListImagesResponses];

export type UploadImageData = {
    body: {
        image: Blob | File;
    };
    path?: never;
    query?: never;
    url: '/images';
};

export type UploadImageErrors = {
    /**
     * No image provided
     */
    400: Error;
};

export type UploadImageError = UploadImageErrors[keyof UploadImageErrors];

export type UploadImageResponses = {
    /**
     * Image uploaded successfully
     */
    201: ImageName;
};

export type UploadImageResponse = UploadImageResponses[keyof UploadImageResponses];

export type DeleteImageData = {
    body?: never;
    path: {
        /**
         * Image filename
         */
        name: string;
    };
    query?: never;
    url: '/images/{name}';
};

export type DeleteImageErrors = {
    /**
     * Image not found
     */
    404: Error;
};

export type DeleteImageError = DeleteImageErrors[keyof DeleteImageErrors];

export type DeleteImageResponses = {
    /**
     * Image deleted
     */
    200: ImageDeleted;
};

export type DeleteImageResponse = DeleteImageResponses[keyof DeleteImageResponses];

export type GetImageData = {
    body?: never;
    path: {
        /**
         * Image filename
         */
        name: string;
    };
    query?: never;
    url: '/images/{name}';
};

export type GetImageErrors = {
    /**
     * Image not found
     */
    404: Error;
};

export type GetImageError = GetImageErrors[keyof GetImageErrors];

export type GetImageResponses = {
    /**
     * The image file
     */
    200: Blob | File;
};

export type GetImageResponse = GetImageResponses[keyof GetImageResponses];

export type UpdateImageData = {
    body?: {
        /**
         * New filename for the image
         */
        newName?: string;
    };
    path: {
        /**
         * Image filename
         */
        name: string;
    };
    query?: never;
    url: '/images/{name}';
};

export type UpdateImageErrors = {
    /**
     * No newName or image provided
     */
    400: Error;
    /**
     * Image not found
     */
    404: Error;
    /**
     * Target name already exists
     */
    409: Error;
};

export type UpdateImageError = UpdateImageErrors[keyof UpdateImageErrors];

export type UpdateImageResponses = {
    /**
     * Image updated
     */
    200: ImageName;
};

export type UpdateImageResponse = UpdateImageResponses[keyof UpdateImageResponses];

export type GetImageThumbnailData = {
    body?: never;
    path: {
        /**
         * Image filename
         */
        name: string;
    };
    query?: never;
    url: '/images/{name}/thumbnail';
};

export type GetImageThumbnailErrors = {
    /**
     * Image not found
     */
    404: Error;
};

export type GetImageThumbnailError = GetImageThumbnailErrors[keyof GetImageThumbnailErrors];

export type GetImageThumbnailResponses = {
    /**
     * JPEG thumbnail image
     */
    200: Blob | File;
};

export type GetImageThumbnailResponse = GetImageThumbnailResponses[keyof GetImageThumbnailResponses];