# image_server/openapi.yaml
openapi: 3.0.3
info:
  title: GU Engine Image API
  version: 1.0.0
  description: CRUD API for managing images in the visual novel engine

servers:
  - url: http://localhost:3007

paths:
  /images:
    get:
      operationId: listImages
      summary: List all image names
      responses:
        '200':
          description: Array of image names
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ImageName'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      operationId: uploadImage
      summary: Upload a new image
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - image
              properties:
                image:
                  type: string
                  format: binary
      responses:
        '201':
          description: Image uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageName'
        '400':
          description: No image provided
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /images/{name}:
    parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
        description: Image filename

    get:
      operationId: getImage
      summary: Get full-size image by name
      responses:
        '200':
          description: The image file
          content:
            image/*:
              schema:
                type: string
                format: binary
        '404':
          description: Image not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      operationId: deleteImage
      summary: Delete an image
      responses:
        '200':
          description: Image deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageDeleted'
        '404':
          description: Image not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      operationId: updateImage
      summary: Rename image or replace its file
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: string
                  format: binary
                  description: New image file to replace the existing one
          application/json:
            schema:
              type: object
              properties:
                newName:
                  type: string
                  description: New filename for the image
      responses:
        '200':
          description: Image updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageName'
        '400':
          description: No newName or image provided
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Image not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Target name already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /images/{name}/thumbnail:
    parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
        description: Image filename

    get:
      operationId: getImageThumbnail
      summary: Get resized thumbnail of an image
      responses:
        '200':
          description: JPEG thumbnail image
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
        '404':
          description: Image not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    ImageName:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: Image filename

    ImageDeleted:
      type: object
      required:
        - deleted
      properties:
        deleted:
          type: string
          description: Name of the deleted image

    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message